Developed Client-Side Frontend Interface
The Frontend of the Web App is now complete, it has 3 main components utilized by a singular Index Page, which are: the SearchForm, the DisclaimerSection, and the ExamResults. The Index Page also houses buttons to navigate to the CPU Official Website, and a button to Report Bugs and Issues to a specified email address. Mock Data was used to test out the display capabilities of the ExamResults component, which will soon be wired up to be served by the exam-schedule Endpoint in our Express Server.